# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Service",
        "description": "API by for longnguyen",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/hairdresser": {
            "get": {
                "tags": [
                    "Hairdresser"
                ],
                "summary": "Get List Hairdresser",
                "description": "Display a listing of the resource.",
                "operationId": "getHairdressers",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "An error occurred while Display a listing the hairdresser."
                    }
                }
            },
            "post": {
                "tags": [
                    "Hairdresser"
                ],
                "summary": "Create Hairdresser",
                "description": "Store a newly created resource in storage.",
                "operationId": "createHairdresser",
                "requestBody": {
                    "description": "Create Hairdresser",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "hairdresser_name",
                                    "hairdresser_phone",
                                    "hairdresser_email",
                                    "hairdresser_images",
                                    "hairdresser_status"
                                ],
                                "properties": {
                                    "hairdresser_name": {
                                        "type": "string",
                                        "example": "Nguyen Van A"
                                    },
                                    "hairdresser_phone": {
                                        "type": "string",
                                        "example": "0123456789"
                                    },
                                    "hairdresser_email": {
                                        "type": "string",
                                        "example": "example@email.com"
                                    },
                                    "hairdresser_images": {
                                        "type": "string",
                                        "format": "binary",
                                        "example": ""
                                    },
                                    "hairdresser_status": {
                                        "type": "integer",
                                        "format": "int64",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "422": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "An error occurred while creating the Hairdresser."
                    }
                }
            }
        },
        "/api/v1/hairdresser/{id}": {
            "delete": {
                "tags": [
                    "Hairdresser"
                ],
                "summary": "Delete Hairdresser",
                "description": "Delete Hairdresser",
                "operationId": "deleteHairdresser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of hairdresser to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hairdresser deleted successfully"
                    },
                    "500": {
                        "description": "An error occurred while deleting the hairdresser."
                    }
                }
            },
            "patch": {
                "tags": [
                    "Hairdresser"
                ],
                "summary": "Update Hairdresser",
                "description": "Update the specified resource in storage.",
                "operationId": "updateHairdresser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of hairdresser that needs to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update Hairdresser",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "hairdresser_name",
                                    "hairdresser_phone",
                                    "hairdresser_email",
                                    "hairdresser_images",
                                    "hairdresser_status"
                                ],
                                "properties": {
                                    "hairdresser_name": {
                                        "type": "string",
                                        "example": "Nguyen Van A"
                                    },
                                    "hairdresser_phone": {
                                        "type": "string",
                                        "example": "0123456789"
                                    },
                                    "hairdresser_email": {
                                        "type": "string",
                                        "example": "example@email.com"
                                    },
                                    "hairdresser_images": {
                                        "type": "string",
                                        "format": "binary",
                                        "example": ""
                                    },
                                    "hairdresser_status": {
                                        "type": "integer",
                                        "format": "int64",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "An error occurred while updating the hairdresser."
                    }
                }
            }
        },
...
    }
}